Fix silent error on modification of FAST input
For example, if we open a FAST file and try to change a variable, like
obj.turbineFile['CompElast'] = 0
and that variable does not exist, it would silently skip the setting of
such variable. That would prevent typos or mistakes from being properly
caught, such as 'CompSeaState', rather than the correct 'CompSeaSt' for
the SeaState module.